+++ src/ProjectInfo.cs
@@ -5,10 +5,11 @@
 // Copyright (c) A Pretty Cool Program. All rights reserved.
 // Licensed under the Apache 2.0 license.
 // -----------------------------------------------------------------------------
-// Version 0.9.20.0
+// Version 0.9.26.x
 // =============================================================================
 
 namespace TingenTransmorger;
+
 internal class ProjectInfo
 {
     // This class is only used for informational purposes, and does not contain executable code.